git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@33140 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/bus_managers/acpi/dispatcher/dswstate.c
@@ -1,7 +1,6 @@
 /******************************************************************************
  *
  * Module Name: dswstate - Dispatcher parse tree walk management routines
- *              $Revision: 1.105 $
  *
  *****************************************************************************/
 
@@ -9,7 +8,7 @@
  *
  * 1. Copyright Notice
  *
- * Some or all of this work - Copyright (c) 1999 - 2008, Intel Corp.
+ * Some or all of this work - Copyright (c) 1999 - 2009, Intel Corp.
  * All rights reserved.
  *
  * 2. License
@@ -118,6 +117,7 @@
 #define __DSWSTATE_C__
 
 #include "acpi.h"
+#include "accommon.h"
 #include "acparser.h"
 #include "acdispat.h"
 #include "acnamesp.h"
@@ -189,7 +189,7 @@ AcpiDsResultPop (
     /* Return object of the top element and clean that top element result stack */
 
     WalkState->ResultCount--;
-    Index = WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
+    Index = (UINT32) WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
 
     *Object = State->Results.ObjDesc [Index];
     if (!*Object)
@@ -284,7 +284,7 @@ AcpiDsResultPush (
 
     /* Assign the address of object to the top free element of result stack */
 
-    Index = WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
+    Index = (UINT32) WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
     State->Results.ObjDesc [Index] = Object;
     WalkState->ResultCount++;
